server-side/filtering-api/requirements.txt
aws-cdk.core
aws-cdk.aws-lambda
aws-cdk.aws-apigateway
